Add Backup Vault functionality and other improvements

+++ app/src/main/java/app/ninesevennine/twofactorauthenticator/features/otp/OtpParser.kt
@@ -60,7 +60,7 @@ fun otpParser(url: String): VaultItem? {
     }
 
     return VaultItem(
-        id = Random.nextInt(),
+        id = Random.nextLong(),
         name = accountName ?: "",
         issuer = issuer,
         note = "",
